+++ c_poi.h
@@ -9,15 +9,15 @@ class C_poi
 private:
     QString sCategorie;
     QString sNom;
-    QHash<float, float> hPoint;
+    QHash<double, double> hPoint;
 public:
     C_poi();
     void setCat(QString);
     void setNom(QString);
-    void setPoint(QHash<float,float>);
+    void setPoint(QHash<double,double>);
     QString getCat();
     QString getNom();
-    QHash<float,float> getPoint();
+    QHash<double,double> getPoint();
 };
 
 #endif // C_POI_H
